Support for global Make variables in soong [1/2]
Add files that should be scanned for variables to the makevarsFiles
array in makevar_adapter.go

Those files will automatically be scanned for variables and found
variables will be resolved to a limited extent (in case the values
also contain variables).

This way you can use global make variables in Go.

Change-Id: Ib8368856c8a8b5194d540ff2350e4eecc4414cde
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>