Do not propagate uses lib tags of framework
framework is a bootclasspath library, thus its uses library tags
collected from its dependencies should not be propagated to its reverse
dependencies. This change utilizes the fact that `sdkLibTag` is only
used to add "framework" and "res" as dependencies for java modules that
set `sdk_version` (i.e. sdk_version "private"), and ensures that uses
lib tags are not collected for `sdkLibTag` dependencies.

Test: CI && ABTD
Bug: 383045647
Change-Id: Idcd9f8d8b131a7ceb28b505fe865d5ac74d51a01
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>